+++ CHANGELOG.md
@@ -1,14 +1,16 @@
 # Changelog
 
-## [v2.2.dev3]
+## [v2.2.3.dev0]
 
 ### Added
 - [Core] Cloudpickle 1.6.0 for python>=3.8
 - [Core] Multiprocessing API
-- [Core] --executor option in tests
+- [Core] Added "--mode" option in tests
+- [Core] Sync standalone function logs into a local file
 
 ### Changed
 - [Core] Improved localhost/standalone logging
+- [Core] Allowed to run lithops without configuration
 
 ### Fixed
 - [Core] Fixed some issues in localhost executor logic